pyproject.toml
[tool.poetry]
name = "cars-api"
version = "0.1.0"
description = ""
authors = ["'Kamran <kamrankek48@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
djoser = "^2.2.3"
celery = "^5.4.0"
gevent = "^24.2.1"
bcrypt = "4.2.0"
redis = "^5.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
